embedding-service/requirements.txt
requests
boto3
langchain_ollama
langchain_community
langchain_text_splitters
faiss-cpu
python-dotenv
fastapi
uvicorn
sentence-transformers